+++ system-module/src/main/java/org/module/controller/system/page/SmPageGroupController.java
@@ -1,13 +1,17 @@
 package org.module.controller.system.page;
 
 import org.module.dto.system.page.SmPageGroupDto;
+import org.module.factory.UserFactory;
+import org.module.helper.commons.CodeHelper;
 import org.module.model.system.page.SmPageGroup;
+import org.module.model.system.user.SmUser;
 import org.module.result.EntityResult;
 import org.module.result.PageResult;
 import org.module.result.RootResult;
 import org.module.service.system.page.ISmPageGroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
@@ -20,7 +24,7 @@ public class SmPageGroupController {
 
 	@RequestMapping("index")
 	public String index() {
-		return "jsp/system/page/group";
+		return "jsp/system/page/group/index";
 	}
 
 	@RequestMapping("data")
@@ -29,23 +33,64 @@ public class SmPageGroupController {
 		return service.findEntityToPage(dto);
 	}
 
+	@RequestMapping("addindex")
+	public String addIndex() {
+		return "jsp/system/page/group/add";
+	}
+
 	@RequestMapping("add")
 	@ResponseBody
-	public EntityResult add(SmPageGroup entity) {
-		entity.setCreateUser("add");
-		return service.insertSelective(entity);
+	public RootResult add(SmPageGroup entity) {
+		RootResult result = new RootResult();
+		SmUser user = UserFactory.instance().userInfo();
+		if (user != null) {
+			entity.setCode(CodeHelper.getUniqueCode("SPG"));
+			entity.setCreateUser(user.getCode());
+			result = service.insertSelective(entity);
+		} else {
+			result.setCode(-1);
+			result.setMessage("用户尚未登录");
+		}
+		return result;
+	}
+
+	@RequestMapping("editindex")
+	public String editIndex(String code, ModelMap model) {
+		EntityResult result = service.selectByCode(code);
+		if (result.getCode() == 0) {
+			model.addAttribute("group", result.getEntity());
+			return "jsp/system/page/group/edit";
+		} else {
+			return "jsp/error/404";
+		}
 	}
 
 	@RequestMapping("edit")
 	@ResponseBody
-	public EntityResult edit(SmPageGroup entity) {
-		entity.setUpdateUser("edit");
-		return service.updateByCode(entity);
+	public RootResult edit(SmPageGroup entity) {
+		RootResult result = new RootResult();
+		SmUser user = UserFactory.instance().userInfo();
+		if (user != null) {
+			entity.setUpdateUser(user.getCode());
+			result = service.updateByCode(entity);
+		} else {
+			result.setCode(-1);
+			result.setMessage("用户尚未登录");
+		}
+		return result;
 	}
 
 	@RequestMapping("del")
 	@ResponseBody
 	public RootResult del(String code) {
-		return service.deleteByCode(code);
+		RootResult result = new RootResult();
+		SmUser user = UserFactory.instance().userInfo();
+		if (user != null) {
+			result = service.deleteByCode(code);
+		} else {
+			result.setCode(-1);
+			result.setMessage("用户尚未登录");
+		}
+		return result;
 	}
 }
